automatic version generation workflow
When requested, this workflow will collect all pull requests since the
last release (or all pull requests in the repository), and determine the
version bump based on the configuration in `.github/release.yml`. If any
of the pull requests match labels within the changelog categories, the
category is configured to bump a version, and an existing version bump
does not already have a higher precedence, the type of version to bump
will be changed. If the version is set to be bumped, the workflow then
runs `cargo bump` to update `Cargo.toml` and create a Git tag/commit for
the release. Finally, it runs `git push --follow-tags` to ensure that
any changes are pushed back to GitHub for publishing.